requirements.txt
# Core ML and AI frameworks (pipeline venv uses cu124)
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124

# Document processing
docling==2.58.0
pymupdf==1.24.4
pymupdf4llm==0.0.5
pdfplumber==0.11.0
pytesseract==0.3.13
Pillow==10.4.0

# Transformers and specialized models
transformers==4.46.2
sentence-transformers==3.0.1
accelerate>=1.0.0,<2
bitsandbytes==0.43.1
FlagEmbedding==1.2.9

# LangChain and AI orchestration
langchain==0.3.7
langchain-community==0.3.5
langchain-ollama==0.2.0
langgraph==0.2.59

# Data processing and validation
pydantic==2.9.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1

# Enhanced NLP and topic modeling
bertopic==0.16.4
keybert==0.8.5
vadersentiment==3.3.2

# JSON repair and validation
json-repair==0.30.2
PyYAML==6.0.2
httpx==0.27.2
beautifulsoup4==4.12.3

# Database and storage
neo4j==5.25.0
chromadb==0.5.15
qdrant-client==1.12.0
rank-bm25==0.2.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.10

# Monitoring and utilities
pynvml==11.5.3
psutil==5.9.8
tqdm==4.66.4
structlog==24.1.0
tenacity==8.2.3
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation==0.46b0
langfuse==2.39.1
phoenix==0.9.0
sentry-sdk==1.39.2

# CLI and interface
click>=8.2.0,<9
rich==13.7.1
streamlit==1.39.0

# Graph and knowledge modelling
networkx==3.3

# Google analytics/search integrations
google-analytics-data==0.18.5
google-api-python-client==2.147.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Development and testing
pytest==8.2.2
black==24.4.2
flake8==7.0.0
openai>=1.65.2,<2
anthropic==0.46.0
